Add simulation execution neuron directory
