+++ obi_one/scientific/tasks/simulation_execution/neuron/__init__.py
@@ -1,10 +1,16 @@
-from obi_one.scientific.tasks.simulation_execution.neuron import (
+from obi_one.scientific.tasks.simulation_execution.neuron.circuit_simulation_execution import (
     CircuitSimulationExecutionSingleConfig,
     CircuitSimulationExecutionTask,
+)
+from obi_one.scientific.tasks.simulation_execution.neuron.ion_channel_simulation_execution import (
     IonChannelModelSimulationExecutionSingleConfig,
     IonChannelModelSimulationExecutionTask,
+)
+from obi_one.scientific.tasks.simulation_execution.neuron.single_neuron_simulation_execution import (  # noqa: E501
     SingleNeuronSimulationExecutionSingleConfig,
     SingleNeuronSimulationExecutionTask,
+)
+from obi_one.scientific.tasks.simulation_execution.neuron.single_neuron_synaptome_simulation_execution import (  # noqa: E501
     SingleNeuronSynaptomeSimulationExecutionSingleConfig,
     SingleNeuronSynaptomeSimulationExecutionTask,
 )
