+++ tests/obi_one/scientific/tasks/simulation_execution/neuron/test_ion_channel_simulation_execution.py
@@ -2,29 +2,34 @@ from unittest.mock import MagicMock, patch
 from uuid import UUID
 
 import pytest
-from obi_one.scientific.library.simulation.schemas import (
+
+from obi_one.core.info import Info
+from obi_one.scientific.blocks.ion_channel_model.ion_channel_model import (
+    IonChannelModelWithConductance,
+)
+from obi_one.scientific.from_id.ion_channel_model_from_id import IonChannelModelFromID
+from obi_one.scientific.library.simulation.neuron.schemas import (
     NeurodamusMechanismBuild,
     NeurodamusSimulationParameters,
     SimulationResults,
 )
-
-from obi_one.core.info import Info
-from obi_one.scientific.from_id.memodel_from_id import MEModelFromID
-from obi_one.scientific.tasks.generate_simulations.config.neuron.neuron_me_model import (
-    MEModelSimulationScanConfig,
-    MEModelSimulationSingleConfig,
+from obi_one.scientific.tasks.generate_simulations.config.neuron.neuron_ion_channel_models import (
+    IonChannelModelSimulationScanConfig,
+    IonChannelModelSimulationSingleConfig,
 )
-from obi_one.scientific.tasks.simulation_execution import (
-    single_neuron_simulation_execution as test_module,
+from obi_one.scientific.tasks.simulation_execution.neuron import (
+    ion_channel_simulation_execution as test_module,
 )
-from obi_one.scientific.tasks.simulation_execution.single_neuron_simulation_execution import (
-    SingleNeuronSimulationExecutionSingleConfig,
-    SingleNeuronSimulationExecutionTask,
+from obi_one.scientific.tasks.simulation_execution.neuron.ion_channel_simulation_execution import (
+    IonChannelModelSimulationExecutionSingleConfig,
+    IonChannelModelSimulationExecutionTask,
 )
 from obi_one.types import SimulationBackend
 
-_BASE = "obi_one.scientific.tasks.simulation_execution.base"
-_SINGLE_NEURON = "obi_one.scientific.tasks.simulation_execution.single_neuron_simulation_execution"
+_BASE = "obi_one.scientific.tasks.simulation_execution.neuron.base"
+_ION_CHANNEL = (
+    "obi_one.scientific.tasks.simulation_execution.neuron.ion_channel_simulation_execution"
+)
 
 
 @pytest.fixture
@@ -36,11 +41,15 @@ def simulation_entity():
 
 @pytest.fixture
 def generation_config(tmp_path, simulation_entity):
-    config = MEModelSimulationSingleConfig(
+    config = IonChannelModelSimulationSingleConfig(
         info=Info(campaign_name="test", campaign_description="test"),
-        initialize=MEModelSimulationScanConfig.Initialize(
-            circuit=MEModelFromID(id_str="memodel-id"),
-        ),
+        initialize=IonChannelModelSimulationScanConfig.Initialize(),
+        ion_channel_models={
+            "ic1": IonChannelModelWithConductance(
+                ion_channel_model=IonChannelModelFromID(id_str="test-id"),
+                conductance=1.0,
+            )
+        },
         idx=0,
         scan_output_root=tmp_path,
         coordinate_output_root=tmp_path / "coord",
@@ -51,13 +60,13 @@ def generation_config(tmp_path, simulation_entity):
 
 @pytest.fixture
 def config(tmp_path, simulation_entity):
-    task_config = SingleNeuronSimulationExecutionSingleConfig(
+    config = IonChannelModelSimulationExecutionSingleConfig(
         idx=0,
         scan_output_root=tmp_path,
         coordinate_output_root=tmp_path / "coord",
     )
-    task_config.set_single_entity(simulation_entity)
-    return task_config
+    config.set_single_entity(simulation_entity)
+    return config
 
 
 @pytest.fixture
@@ -79,12 +88,21 @@ def _neurodamus_mechanism_build(tmp_path):
     )
 
 
-@patch(f"{_SINGLE_NEURON}.SingleNeuronSimulationExecutionTask.get_generation_single_config")
+def _neurodamus_simulation_parameters(tmp_path, mechanism_build):
+    return NeurodamusSimulationParameters(
+        number_of_cells=1,
+        stop_time=100.0,
+        config_file=tmp_path / "config.json",
+        mechanism_build=mechanism_build,
+    )
+
+
+@patch(f"{_ION_CHANNEL}.IonChannelModelSimulationExecutionTask.get_generation_single_config")
 @patch(f"{_BASE}.run_simulation")
 @patch(f"{_BASE}.get_simulation_parameters")
 @patch(f"{_BASE}.stage_simulation")
 @patch(f"{_BASE}.compile_mechanisms")
-@patch(f"{_SINGLE_NEURON}.stage_memodel_as_circuit")
+@patch(f"{_ION_CHANNEL}.stage_ion_channel_models_as_circuit")
 @patch(f"{_BASE}.create_dir")
 def test_execute_local_does_not_register(
     mock_create_dir,
@@ -109,18 +127,13 @@ def test_execute_local_does_not_register(
     mechanism_build = _neurodamus_mechanism_build(tmp_path)
     mock_compile.return_value = mechanism_build
     mock_stage_sim.return_value = tmp_path / "sim_config.json"
-    mock_get_params.return_value = NeurodamusSimulationParameters(
-        number_of_cells=1,
-        stop_time=100.0,
-        config_file=tmp_path / "config.json",
-        mechanism_build=mechanism_build,
-    )
+    mock_get_params.return_value = _neurodamus_simulation_parameters(tmp_path, mechanism_build)
     mock_run_simulation.return_value = SimulationResults(
         spike_report_file=tmp_path / "spikes.h5",
         voltage_report_files=[],
     )
 
-    task = SingleNeuronSimulationExecutionTask(config=config)
+    task = IonChannelModelSimulationExecutionTask(config=config)
     task.execute(db_client=db_client, execution_activity_id=None)
 
     mock_stage_circuit.assert_called_once()
@@ -130,16 +143,19 @@ def test_execute_local_does_not_register(
         simulation_backend=SimulationBackend.neurodamus,
     )
     db_client.get_entity.assert_not_called()
+    mock_stage_sim.assert_called_once()
+    mock_get_params.assert_called_once()
+    mock_run_simulation.assert_called_once()
     db_client.update_entity.assert_not_called()
 
 
-@patch(f"{_SINGLE_NEURON}.SingleNeuronSimulationExecutionTask.get_generation_single_config")
+@patch(f"{_ION_CHANNEL}.IonChannelModelSimulationExecutionTask.get_generation_single_config")
 @patch(f"{_BASE}.register_simulation_results")
 @patch(f"{_BASE}.run_simulation")
 @patch(f"{_BASE}.get_simulation_parameters")
 @patch(f"{_BASE}.stage_simulation")
 @patch(f"{_BASE}.compile_mechanisms")
-@patch(f"{_SINGLE_NEURON}.stage_memodel_as_circuit")
+@patch(f"{_ION_CHANNEL}.stage_ion_channel_models_as_circuit")
 @patch(f"{_BASE}.create_dir")
 def test_execute_tracked_registers_and_updates_activity(
     mock_create_dir,
@@ -168,57 +184,58 @@ def test_execute_tracked_registers_and_updates_activity(
     execution_activity.id = "act-456"
     db_client.get_entity.return_value = execution_activity
     mock_stage_sim.return_value = tmp_path / "sim_config.json"
+    mock_get_params.return_value = _neurodamus_simulation_parameters(tmp_path, mechanism_build)
     sim_results = SimulationResults(
         spike_report_file=tmp_path / "spikes.h5",
         voltage_report_files=[],
-    )
-    mock_get_params.return_value = NeurodamusSimulationParameters(
-        number_of_cells=1,
-        stop_time=100.0,
-        config_file=tmp_path / "config.json",
-        mechanism_build=mechanism_build,
     )
     mock_run_simulation.return_value = sim_results
     generated_entity = MagicMock()
     generated_entity.id = "gen-789"
     mock_register.return_value = generated_entity
 
-    task = SingleNeuronSimulationExecutionTask(config=config)
+    task = IonChannelModelSimulationExecutionTask(config=config)
     task.execute(db_client=db_client, execution_activity_id="act-456")
 
     db_client.get_entity.assert_called_once_with(
         entity_id="act-456",
-        entity_type=test_module.SingleNeuronSimulationExecutionTask.activity_type,
+        entity_type=test_module.IonChannelModelSimulationExecutionTask.activity_type,
     )
     mock_register.assert_called_once()
+    call_kw = mock_register.call_args[1]
+    assert call_kw["client"] == db_client
+    assert call_kw["simulation_results"] == sim_results
+    assert call_kw["simulation_metadata"].simulation_id == config.single_entity.id
     db_client.update_entity.assert_called_once_with(
         entity_id="act-456",
-        entity_type=test_module.SingleNeuronSimulationExecutionTask.activity_type,
+        entity_type=test_module.IonChannelModelSimulationExecutionTask.activity_type,
         attrs_or_entity={"generated_ids": ["gen-789"]},
     )
 
 
-@patch(f"{_SINGLE_NEURON}.deserialize_obi_object_from_json_data")
-@patch(f"{_SINGLE_NEURON}.db_sdk.select_json_asset_content")
+@patch(f"{_ION_CHANNEL}.deserialize_obi_object_from_json_data")
+@patch(f"{_ION_CHANNEL}.db_sdk.select_json_asset_content")
 def test_get_generation_single_config(
     mock_select_json, mock_deserialize, config, generation_config
 ):
-    mock_select_json.return_value = {"type": "memodel"}
+    mock_select_json.return_value = {"type": "ion_channel"}
     mock_deserialize.return_value = generation_config
 
-    task = SingleNeuronSimulationExecutionTask(config=config)
+    task = IonChannelModelSimulationExecutionTask(config=config)
     result = task.get_generation_single_config(db_client=MagicMock())
 
     assert result is generation_config
     mock_select_json.assert_called_once()
 
 
-@patch(f"{_SINGLE_NEURON}.deserialize_obi_object_from_json_data")
-@patch(f"{_SINGLE_NEURON}.db_sdk.select_json_asset_content")
+@patch(f"{_ION_CHANNEL}.deserialize_obi_object_from_json_data")
+@patch(f"{_ION_CHANNEL}.db_sdk.select_json_asset_content")
 def test_get_generation_single_config_wrong_type(mock_select_json, mock_deserialize, config):
     mock_select_json.return_value = {"type": "other"}
     mock_deserialize.return_value = MagicMock()
 
-    task = SingleNeuronSimulationExecutionTask(config=config)
-    with pytest.raises(test_module.OBIONEError, match="Expected MEModelSimulationSingleConfig"):
+    task = IonChannelModelSimulationExecutionTask(config=config)
+    with pytest.raises(
+        test_module.OBIONEError, match="Expected IonChannelModelSimulationSingleConfig"
+    ):
         task.get_generation_single_config(db_client=MagicMock())
