+++ tests/obi_one/scientific/tasks/simulation_execution/neuron/test_single_neuron_simulation_execution.py
@@ -2,32 +2,31 @@ from unittest.mock import MagicMock, patch
 from uuid import UUID
 
 import pytest
-from obi_one.scientific.library.simulation.schemas import (
+
+from obi_one.core.info import Info
+from obi_one.scientific.from_id.memodel_from_id import MEModelFromID
+from obi_one.scientific.library.simulation.neuron.schemas import (
     NeurodamusMechanismBuild,
     NeurodamusSimulationParameters,
     SimulationResults,
 )
-
-from obi_one.core.info import Info
-from obi_one.scientific.blocks.ion_channel_model.ion_channel_model import (
-    IonChannelModelWithConductance,
+from obi_one.scientific.tasks.generate_simulations.config.neuron.neuron_me_model import (
+    MEModelSimulationScanConfig,
+    MEModelSimulationSingleConfig,
 )
-from obi_one.scientific.from_id.ion_channel_model_from_id import IonChannelModelFromID
-from obi_one.scientific.tasks.generate_simulations.config.neuron.neuron_ion_channel_models import (
-    IonChannelModelSimulationScanConfig,
-    IonChannelModelSimulationSingleConfig,
+from obi_one.scientific.tasks.simulation_execution.neuron import (
+    single_neuron_simulation_execution as test_module,
 )
-from obi_one.scientific.tasks.simulation_execution import (
-    ion_channel_simulation_execution as test_module,
-)
-from obi_one.scientific.tasks.simulation_execution.ion_channel_simulation_execution import (
-    IonChannelModelSimulationExecutionSingleConfig,
-    IonChannelModelSimulationExecutionTask,
+from obi_one.scientific.tasks.simulation_execution.neuron.single_neuron_simulation_execution import (  # noqa: E501
+    SingleNeuronSimulationExecutionSingleConfig,
+    SingleNeuronSimulationExecutionTask,
 )
 from obi_one.types import SimulationBackend
 
-_BASE = "obi_one.scientific.tasks.simulation_execution.base"
-_ION_CHANNEL = "obi_one.scientific.tasks.simulation_execution.ion_channel_simulation_execution"
+_BASE = "obi_one.scientific.tasks.simulation_execution.neuron.base"
+_SINGLE_NEURON = (
+    "obi_one.scientific.tasks.simulation_execution.neuron.single_neuron_simulation_execution"
+)
 
 
 @pytest.fixture
@@ -39,15 +38,11 @@ def simulation_entity():
 
 @pytest.fixture
 def generation_config(tmp_path, simulation_entity):
-    config = IonChannelModelSimulationSingleConfig(
+    config = MEModelSimulationSingleConfig(
         info=Info(campaign_name="test", campaign_description="test"),
-        initialize=IonChannelModelSimulationScanConfig.Initialize(),
-        ion_channel_models={
-            "ic1": IonChannelModelWithConductance(
-                ion_channel_model=IonChannelModelFromID(id_str="test-id"),
-                conductance=1.0,
-            )
-        },
+        initialize=MEModelSimulationScanConfig.Initialize(
+            circuit=MEModelFromID(id_str="memodel-id"),
+        ),
         idx=0,
         scan_output_root=tmp_path,
         coordinate_output_root=tmp_path / "coord",
@@ -58,13 +53,13 @@ def generation_config(tmp_path, simulation_entity):
 
 @pytest.fixture
 def config(tmp_path, simulation_entity):
-    config = IonChannelModelSimulationExecutionSingleConfig(
+    task_config = SingleNeuronSimulationExecutionSingleConfig(
         idx=0,
         scan_output_root=tmp_path,
         coordinate_output_root=tmp_path / "coord",
     )
-    config.set_single_entity(simulation_entity)
-    return config
+    task_config.set_single_entity(simulation_entity)
+    return task_config
 
 
 @pytest.fixture
@@ -86,21 +81,12 @@ def _neurodamus_mechanism_build(tmp_path):
     )
 
 
-def _neurodamus_simulation_parameters(tmp_path, mechanism_build):
-    return NeurodamusSimulationParameters(
-        number_of_cells=1,
-        stop_time=100.0,
-        config_file=tmp_path / "config.json",
-        mechanism_build=mechanism_build,
-    )
-
-
-@patch(f"{_ION_CHANNEL}.IonChannelModelSimulationExecutionTask.get_generation_single_config")
+@patch(f"{_SINGLE_NEURON}.SingleNeuronSimulationExecutionTask.get_generation_single_config")
 @patch(f"{_BASE}.run_simulation")
 @patch(f"{_BASE}.get_simulation_parameters")
 @patch(f"{_BASE}.stage_simulation")
 @patch(f"{_BASE}.compile_mechanisms")
-@patch(f"{_ION_CHANNEL}.stage_ion_channel_models_as_circuit")
+@patch(f"{_SINGLE_NEURON}.stage_memodel_as_circuit")
 @patch(f"{_BASE}.create_dir")
 def test_execute_local_does_not_register(
     mock_create_dir,
@@ -125,13 +111,18 @@ def test_execute_local_does_not_register(
     mechanism_build = _neurodamus_mechanism_build(tmp_path)
     mock_compile.return_value = mechanism_build
     mock_stage_sim.return_value = tmp_path / "sim_config.json"
-    mock_get_params.return_value = _neurodamus_simulation_parameters(tmp_path, mechanism_build)
+    mock_get_params.return_value = NeurodamusSimulationParameters(
+        number_of_cells=1,
+        stop_time=100.0,
+        config_file=tmp_path / "config.json",
+        mechanism_build=mechanism_build,
+    )
     mock_run_simulation.return_value = SimulationResults(
         spike_report_file=tmp_path / "spikes.h5",
         voltage_report_files=[],
     )
 
-    task = IonChannelModelSimulationExecutionTask(config=config)
+    task = SingleNeuronSimulationExecutionTask(config=config)
     task.execute(db_client=db_client, execution_activity_id=None)
 
     mock_stage_circuit.assert_called_once()
@@ -141,19 +132,16 @@ def test_execute_local_does_not_register(
         simulation_backend=SimulationBackend.neurodamus,
     )
     db_client.get_entity.assert_not_called()
-    mock_stage_sim.assert_called_once()
-    mock_get_params.assert_called_once()
-    mock_run_simulation.assert_called_once()
     db_client.update_entity.assert_not_called()
 
 
-@patch(f"{_ION_CHANNEL}.IonChannelModelSimulationExecutionTask.get_generation_single_config")
+@patch(f"{_SINGLE_NEURON}.SingleNeuronSimulationExecutionTask.get_generation_single_config")
 @patch(f"{_BASE}.register_simulation_results")
 @patch(f"{_BASE}.run_simulation")
 @patch(f"{_BASE}.get_simulation_parameters")
 @patch(f"{_BASE}.stage_simulation")
 @patch(f"{_BASE}.compile_mechanisms")
-@patch(f"{_ION_CHANNEL}.stage_ion_channel_models_as_circuit")
+@patch(f"{_SINGLE_NEURON}.stage_memodel_as_circuit")
 @patch(f"{_BASE}.create_dir")
 def test_execute_tracked_registers_and_updates_activity(
     mock_create_dir,
@@ -182,58 +170,57 @@ def test_execute_tracked_registers_and_updates_activity(
     execution_activity.id = "act-456"
     db_client.get_entity.return_value = execution_activity
     mock_stage_sim.return_value = tmp_path / "sim_config.json"
-    mock_get_params.return_value = _neurodamus_simulation_parameters(tmp_path, mechanism_build)
     sim_results = SimulationResults(
         spike_report_file=tmp_path / "spikes.h5",
         voltage_report_files=[],
+    )
+    mock_get_params.return_value = NeurodamusSimulationParameters(
+        number_of_cells=1,
+        stop_time=100.0,
+        config_file=tmp_path / "config.json",
+        mechanism_build=mechanism_build,
     )
     mock_run_simulation.return_value = sim_results
     generated_entity = MagicMock()
     generated_entity.id = "gen-789"
     mock_register.return_value = generated_entity
 
-    task = IonChannelModelSimulationExecutionTask(config=config)
+    task = SingleNeuronSimulationExecutionTask(config=config)
     task.execute(db_client=db_client, execution_activity_id="act-456")
 
     db_client.get_entity.assert_called_once_with(
         entity_id="act-456",
-        entity_type=test_module.IonChannelModelSimulationExecutionTask.activity_type,
+        entity_type=test_module.SingleNeuronSimulationExecutionTask.activity_type,
     )
     mock_register.assert_called_once()
-    call_kw = mock_register.call_args[1]
-    assert call_kw["client"] == db_client
-    assert call_kw["simulation_results"] == sim_results
-    assert call_kw["simulation_metadata"].simulation_id == config.single_entity.id
     db_client.update_entity.assert_called_once_with(
         entity_id="act-456",
-        entity_type=test_module.IonChannelModelSimulationExecutionTask.activity_type,
+        entity_type=test_module.SingleNeuronSimulationExecutionTask.activity_type,
         attrs_or_entity={"generated_ids": ["gen-789"]},
     )
 
 
-@patch(f"{_ION_CHANNEL}.deserialize_obi_object_from_json_data")
-@patch(f"{_ION_CHANNEL}.db_sdk.select_json_asset_content")
+@patch(f"{_SINGLE_NEURON}.deserialize_obi_object_from_json_data")
+@patch(f"{_SINGLE_NEURON}.db_sdk.select_json_asset_content")
 def test_get_generation_single_config(
     mock_select_json, mock_deserialize, config, generation_config
 ):
-    mock_select_json.return_value = {"type": "ion_channel"}
+    mock_select_json.return_value = {"type": "memodel"}
     mock_deserialize.return_value = generation_config
 
-    task = IonChannelModelSimulationExecutionTask(config=config)
+    task = SingleNeuronSimulationExecutionTask(config=config)
     result = task.get_generation_single_config(db_client=MagicMock())
 
     assert result is generation_config
     mock_select_json.assert_called_once()
 
 
-@patch(f"{_ION_CHANNEL}.deserialize_obi_object_from_json_data")
-@patch(f"{_ION_CHANNEL}.db_sdk.select_json_asset_content")
+@patch(f"{_SINGLE_NEURON}.deserialize_obi_object_from_json_data")
+@patch(f"{_SINGLE_NEURON}.db_sdk.select_json_asset_content")
 def test_get_generation_single_config_wrong_type(mock_select_json, mock_deserialize, config):
     mock_select_json.return_value = {"type": "other"}
     mock_deserialize.return_value = MagicMock()
 
-    task = IonChannelModelSimulationExecutionTask(config=config)
-    with pytest.raises(
-        test_module.OBIONEError, match="Expected IonChannelModelSimulationSingleConfig"
-    ):
+    task = SingleNeuronSimulationExecutionTask(config=config)
+    with pytest.raises(test_module.OBIONEError, match="Expected MEModelSimulationSingleConfig"):
         task.get_generation_single_config(db_client=MagicMock())
